Refactored long handleCAS function
