+++ commands/register.go
@@ -2,7 +2,6 @@ package commands
 
 import (
 	"errors"
-	"fmt"
 	"net/url"
 
 	"github.com/albshin/tutescrew/config"
@@ -17,7 +16,7 @@ func (r *Register) handle(ctx Context) error {
 	// Check if student is already registered
 	ch, err := ctx.Sess.State.Channel(ctx.Msg.ChannelID)
 	if err != nil {
-		fmt.Println("Could not get channel")
+		return err
 	}
 	g, _ := ctx.Sess.State.Guild(ch.GuildID)
 
@@ -27,11 +26,17 @@ func (r *Register) handle(ctx Context) error {
 	}
 
 	// Build the full login URL
-	u, _ := url.Parse(r.Config.CASAuthURL)
+	u, err := url.Parse(r.Config.CASAuthURL)
+	if err != nil {
+		return err
+	}
 	q := u.Query()
 
 	// Encode Discord values into the redirect
-	re, _ := url.Parse(r.Config.CASRedirectURL)
+	re, err := url.Parse(r.Config.CASRedirectURL)
+	if err != nil {
+		return err
+	}
 	reque := re.Query()
 	reque.Add("guild", ch.GuildID)
 	reque.Add("discord_id", ctx.Msg.Author.ID)
@@ -41,7 +46,10 @@ func (r *Register) handle(ctx Context) error {
 	q.Add("service", re.String())
 	u.RawQuery = q.Encode()
 
-	usrch, _ := ctx.Sess.UserChannelCreate(ctx.Msg.Author.ID)
+	usrch, err := ctx.Sess.UserChannelCreate(ctx.Msg.Author.ID)
+	if err != nil {
+		return err
+	}
 	ctx.Sess.ChannelMessageSend(usrch.ID, "Please go to "+u.String()+" to start the registration process.")
 
 	return nil
